flock instead of fchmod in C implementation
On existing kernels, fchmod ends up using chmod, so any FD-level cache
on the open file descriptor is not used. This may be fixed in future
kernels. However, flock does actually use the open existing file
descriptor, and its cache. So use flock to test a working FD-level cache
for now, and use the Rust or Go implementation to check that fchmod
works in the future, when the kernel supports it.

Signed-off-by: Oliver Calder <oliver@calder.dev>